refactor: discard unnecessary packages
